+++ tactics2d/map/parser/parse_xodr.py
@@ -2,7 +2,7 @@
 # Copyright (C) 2024, Tactics2D Authors. Released under the GNU GPLv3.
 # @File: parse_xodr.py
 # @Description: This file defines a class for parsing the OpenDRIVE map format.
-# @Author: Yueyuan Li
+# @Author: Tactics2D Team
 # @Version: 1.0.0